# src/main/java/com/nuevospa/gestiontareas/dto/LoginRequest.java
package com.nuevospa.gestiontareas.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class LoginRequest {
    private String email;
    private String password;
} 